[action] update TxData interface
